+++ src/plugins/data/common/query/filter_manager/dedup_filters.ts
@@ -1,0 +1,49 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { filter, find } from 'lodash';
+import { compareFilters, FilterCompareOptions } from './compare_filters';
+import { Filter } from '../../es_query';
+
+/**
+ * Combine 2 filter collections, removing duplicates
+ *
+ * @param {object} existingFilters - The filters to compare to
+ * @param {object} filters - The filters being added
+ * @param {object} comparatorOptions - Parameters to use for comparison
+ *
+ * @returns {object} An array of filters that were not in existing
+ */
+export const dedupFilters = (
+  existingFilters: Filter[],
+  filters: Filter[],
+  comparatorOptions: FilterCompareOptions = {}
+) => {
+  if (!Array.isArray(filters)) {
+    filters = [filters];
+  }
+
+  return filter(
+    filters,
+    (f: Filter) =>
+      !find(existingFilters, (existingFilter: Filter) =>
+        compareFilters(existingFilter, f, comparatorOptions)
+      )
+  );
+};
