+++ src/plugins/data/public/query/filter_manager/index.ts
@@ -19,8 +19,6 @@
 
 export { FilterManager } from './filter_manager';
 
-export { uniqFilters } from './lib/uniq_filters';
 export { mapAndFlattenFilters } from './lib/map_and_flatten_filters';
 export { onlyDisabledFiltersChanged } from './lib/only_disabled';
 export { generateFilters } from './lib/generate_filters';
-export { compareFilters, COMPARE_ALL_OPTIONS } from './lib/compare_filters';
